requirements.txt
pip==7.1.2
tornado
requests
lxml
arrow
cloudant
PyGithub
PyYAML
